tests/wpt/web-platform-tests/content-security-policy/blink-contrib-2/stylehash-default-src.sub.html

<!DOCTYPE html>
<html>
    <head>
    <title>stylehash allowed from default-src</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    </head>

    <body>
    <p id="p">Test</p>
    <style>p#p { color: green; }</style>
    <script>
    var color = window.getComputedStyle(document.querySelector('#p')).color;
    assert_equals(color, "rgb(0, 128, 0)");
    done();
    </script>

    <div id="log"></div>
    <script async defer src="../support/checkReport.sub.js?reportExists=false"></script>
    </body>
</html>
